Add new tuners for each movement dimension
