+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/tuner/HeadingFFTuner.java
@@ -1,78 +1,97 @@
 package com.acmerobotics.relicrecovery.opmodes.tuner;
 
+import com.acmerobotics.library.dashboard.config.Config;
 import com.acmerobotics.library.localization.Pose2d;
-import com.acmerobotics.relicrecovery.path.Path;
-import com.acmerobotics.relicrecovery.path.PointTurn;
+import com.acmerobotics.relicrecovery.path.PathBuilder;
+import com.acmerobotics.relicrecovery.path.PathFollower;
 import com.acmerobotics.relicrecovery.subsystems.MecanumDrive;
 import com.acmerobotics.relicrecovery.subsystems.Robot;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import java.util.Arrays;
-
-@TeleOp(name = "Heading FF Tuner")
+@Config
+@TeleOp
 public class HeadingFFTuner extends LinearOpMode {
-    public static final double LOWER_BOUND = 0;
-    public static final double UPPER_BOUND = 0.075;
-    public static final double TURN_ANGLE = Math.PI / 2;
+    public static double LOWER_BOUND = 0, UPPER_BOUND = 0.5;
+    public static double ANGLE = Math.PI / 2; // rad
+
+    /**
+     * The search confidence codifies how fast the search window homes in on the solution. This
+     * value should be in the range (0, 1] where 1 indicates a true binary search.
+     */
+    public static double SEARCH_CONFIDENCE = 1;
+
+    public static int TRIALS = 3;
 
     private Robot robot;
-    private volatile double value;
+    private PathFollower pathFollower;
 
     @Override
-    public void runOpMode() {
+    public void runOpMode() throws InterruptedException {
         robot = new Robot(this);
-
         robot.drive.enablePositionEstimation();
-
         robot.start();
+
+        pathFollower = robot.drive.getPathFollower();
+
+        telemetry.log().add("Ready. Make sure to clear PID gains.");
+        telemetry.update();
 
         waitForStart();
 
-        double lower = LOWER_BOUND, upper = UPPER_BOUND;
+        telemetry.clear();
+
+        double lowerBound = LOWER_BOUND, upperBound = UPPER_BOUND;
 
         while (opModeIsActive()) {
-            value = (lower + upper) / 2;
+            double value = (LOWER_BOUND + UPPER_BOUND) / 2;
 
-            int numErrors = 3;
-            double errorSum = 0;
-            for (int i = 0; i < numErrors; i++) {
-                errorSum += testFeedforwardCoefficient(value);
+            double meanError = 0;
+            for (int i = 0; i < TRIALS && opModeIsActive(); i++) {
+                meanError += testCoefficient(value);
             }
-            double error = errorSum / numErrors;
-            if (error > 0) {
-                upper = value;
+            meanError /= TRIALS;
+
+            if (meanError > 0) {
+                // the coefficient is too high; lower the upper bound
+                upperBound = (upperBound - value) * (1 - SEARCH_CONFIDENCE) + value;
             } else {
-                lower = value;
+                // the coefficient is too low; raise the lower bound
+                lowerBound = (value - lowerBound) * SEARCH_CONFIDENCE + lowerBound;
             }
+
+            telemetry.addData("value", value);
+            telemetry.addData("lastError", meanError);
+            telemetry.update();
         }
     }
 
-    private double testFeedforwardCoefficient(double coefficient) {
+    /**
+     * Complete a single motion with the provided coefficient and measure its accuracy
+     * @param coefficient axial K_a coefficient to test
+     * @return the mean error for the movement, positive indicates the coefficient is too high
+     */
+    public double testCoefficient(double coefficient) {
         MecanumDrive.HEADING_PID.a = coefficient;
 
-        // reset heading + pose
         robot.drive.setEstimatedPose(new Pose2d(0, 0, 0));
 
-        // turn
-        Path path = new Path(Arrays.asList(
-                new PointTurn(new Pose2d(0, 0, 0), TURN_ANGLE)
-        ));
-        robot.drive.followPath(path);
-
-        telemetry.addData("headingKa", value);
-        telemetry.update();
+        robot.drive.followPath(new PathBuilder(new Pose2d(0, 0, 0))
+                .turn(ANGLE)
+                .build());
 
         double errorSum = 0;
         int numErrors = 0;
 
         while (opModeIsActive() && robot.drive.isFollowingPath()) {
-            errorSum += robot.drive.getPathFollower().getHeadingError();
+            double rawHeadingError = pathFollower.getHeadingError();
+            // If axial K_a is too high, then the error will be positive in when acceleration is
+            // positive and negative when acceleration is negative. Thus we use the sign of the
+            // acceleration to make sure all the errors have matching signs.
+            errorSum += Math.signum(pathFollower.getPoseAcceleration().heading()) * rawHeadingError;
             numErrors++;
-            sleep(5);
+            sleep(25);
         }
-
-        sleep(1000);
 
         return errorSum / numErrors;
     }
